FIX: Integer -> Long

+++ src/main/java/com/backend/baseball/User/helper/AccountHelper.java
@@ -13,7 +13,7 @@ public interface AccountHelper {
     String login(LoginRequestDTO loginReq, HttpServletRequest req, HttpServletResponse res);
 
     //회원 이메일 조회(아이디 조회)--
-    Integer getMemberId(HttpServletRequest req);
+    Long getMemberId(HttpServletRequest req);
 
     //로그인 여부 확인
     boolean isLoggedIn(HttpServletRequest req);
